HisiPkg: D01: Move bootwrapper to SRAM
Reserving DDR memory is a very tricky thing to do. We always need to make
sure we don't overlap with memory regions the guest uncondionally uses.

However, we need to keep the bootwrapper code resident in memory for CPU
hotplug. So let's just move it into SRAM. There it can stay happily ever
after.

Unfortunately we need to make sure we only copy the bootwrapper to SRAM,
not a full NAND block. So this patch also introduces a bounce buffer for
NAND access.

Signed-off-by: Alexander Graf <agraf@suse.de>
Signed-off-by: Wei Xu <xuwei5@huawei.com>