ArmPlatformPkg/Bds: add back in UpdateFdtPath
The commit listed below removed the ability for the user to set the FDT
device path, meaning that users are no longer able to boot via TFTP
because they have no way of specifying the path to the FDT file.

This patch adds back in the UpdateFdtPath function, but updates it to
check the newly named FdtDevicePath -> FdtDevicePaths variable.

  commit 6332ffb0affadde53e45325d81b084bc9936ef02
  Author: Ronald Cron <Ronald.Cron@arm.com>
  Date:   Thu Feb 26 11:02:39 2015 +0000

    ArmPlatformPkg/Bds: Remove any use of the "Fdt" UEFI variable

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>