ArmPlatformPkg/Bds: remove detection of EFI stub
If the ARM BDS detected that there was an EFI stub on the kernel or
application that was being booted, it forced the user to boot it as an
EFI application and didn't give the user an option to boot it as an FDT
or ATAGS kernel.

It was also very slow because it had to load the image from
disk/wherever to check the image.

This patch removes that check and reverts to the original behavious
where the user is asked if it is an EFI application or not.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>