HACK: SimpleTextInOut: Fix ESC and treat ASCII DEL (0x7f) as backspace
ASCII DEL is sent on most Linux terminals when the backspace key is
pressed, so translate it into a backspace code.

Also, the ESC key wasn't being passed in a way that the EFI Shell was
expecting, so this patch also fixes that.

I refactored ReadKeyStroke() to use a switch statement because the code
was very hard to follow and you couldn't quite tell which bit was
handling which key codes.

In the process, I dropped the handling of extra codes like HOME, END and
function keys because they aren't handled properly anyway.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>