ArmPkg: BdsLib: move initrd above 128MiB boundary
Move the initrd to an address greater than the 128MiB boundary as
requested in the kernel recommendation.

We do this by reusing LINX_FDT_MAX_OFFSET, which is above the 128MiB
boundary.

This patch is based off an older Tiancore.  A version based of the tip
of tree was posted to linaro-uefi for review:

https://lists.linaro.org/pipermail/linaro-uefi/2015-July/001016.html

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>